Use long as migration version

+++ sqldelight-gradle-plugin/src/main/kotlin/app/cash/sqldelight/gradle/VerifyMigrationTask.kt
@@ -191,7 +191,7 @@ abstract class VerifyMigrationTask : SqlDelightWorkerTask() {
     }
 
     private fun checkForGaps() {
-      var lastMigrationVersion: Int? = null
+      var lastMigrationVersion: Long? = null
       environment.forMigrationFiles {
         val actual = it.version
         val expected = lastMigrationVersion?.plus(1) ?: actual
